requirements.txt
pytest
webdriver-manager
selenium
requests
allure-pytest
mysql-connector-python
